Move file diagram-options.model
