+++ projects/fhi-angular-highcharts/src/lib/services/geo-json.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
-import { DiagramSerie } from '../diagram-options.model';
+import { DiagramSerie } from '../fhi-diagram/diagram-options.model';
 import { SeriesMapOptions } from 'highcharts';
 
 @Injectable({
